fix: WebUSB grant access
