+++ packages/kit/src/views/FirmwareUpdate/pages/PageFirmwareUpdateInstall.tsx
@@ -1,14 +1,10 @@
-import { useEffect, useMemo, useRef } from 'react';
+import { useMemo } from 'react';
 
 import { Page } from '@onekeyhq/components';
 import {
   EFirmwareUpdateSteps,
   useFirmwareUpdateStepInfoAtom,
 } from '@onekeyhq/kit-bg/src/states/jotai/atoms';
-import {
-  EAppEventBusNames,
-  appEventBus,
-} from '@onekeyhq/shared/src/eventBus/appEventBus';
 import type {
   EModalFirmwareUpdateRoutes,
   IModalFirmwareUpdateParamList,
@@ -24,7 +20,6 @@ import {
   ForceExtensionUpdatingFromExpandTab,
 } from '../components/FirmwareUpdateExitPrevent';
 import { FirmwareUpdatePageLayout } from '../components/FirmwareUpdatePageLayout';
-import { FirmwareUpdatePromptBootloaderWebDevice } from '../components/FirmwareUpdatePromptBootloaderWebDevice';
 import { FirmwareUpdateWarningMessage } from '../components/FirmwareUpdateWarningMessage';
 
 function PageFirmwareUpdateInstall() {
@@ -35,25 +30,7 @@ function PageFirmwareUpdateInstall() {
   const { result } = route.params;
 
   const navigation = useAppNavigation();
-  const [stepInfo, setStepInfo] = useFirmwareUpdateStepInfoAtom();
-  const previousStepInfo = useRef(stepInfo);
-
-  useEffect(() => {
-    const fn = () => {
-      previousStepInfo.current = stepInfo;
-      setStepInfo({
-        step: EFirmwareUpdateSteps.requestDeviceInBootloaderForWebDevice,
-        payload: undefined,
-      });
-    };
-    appEventBus.on(EAppEventBusNames.RequestDeviceInBootloaderForWebDevice, fn);
-    return () => {
-      appEventBus.off(
-        EAppEventBusNames.RequestDeviceInBootloaderForWebDevice,
-        fn,
-      );
-    };
-  }, [setStepInfo, stepInfo]);
+  const [stepInfo] = useFirmwareUpdateStepInfoAtom();
 
   /*
      await backgroundApiProxy.serviceFirmwareUpdate.startFirmwareUpdateWorkflow(
@@ -72,6 +49,8 @@ function PageFirmwareUpdateInstall() {
     if (
       stepInfo.step === EFirmwareUpdateSteps.updateStart ||
       stepInfo.step === EFirmwareUpdateSteps.installing ||
+      stepInfo.step ===
+        EFirmwareUpdateSteps.requestDeviceInBootloaderForWebDevice ||
       stepInfo.step === EFirmwareUpdateSteps.updateDone
     ) {
       const isDone = stepInfo.step === EFirmwareUpdateSteps.updateDone;
@@ -93,20 +72,6 @@ function PageFirmwareUpdateInstall() {
         navigation.pop();
       });
       return <FirmwareUpdateExitPrevent shouldPreventRemove={false} />;
-    }
-
-    if (
-      stepInfo.step ===
-      EFirmwareUpdateSteps.requestDeviceInBootloaderForWebDevice
-    ) {
-      return (
-        <>
-          <FirmwareUpdatePromptBootloaderWebDevice
-            previousStepInfo={previousStepInfo.current}
-          />
-          <FirmwareUpdateExitPrevent />
-        </>
-      );
     }
 
     return (
